control de permisos para ProductoCnotroller

+++ application/controllers/ProductoController.php
@@ -7,7 +7,8 @@ include_once(dirname(__DIR__).'/models/CategoriaModel.php');
 class ProductoController extends Controller
 {
 
-    public function __construct(){
+    public function __construct()
+    {
         parent::__construct();
         $this->load->model('ProductosModel');
         $this->load->model('CategoriaModel');
@@ -20,46 +21,64 @@ class ProductoController extends Controller
         return (($rol == '0') || ($rol == '1'));
     }
 
-    public function productosListar(){
-        $this->paginaCorrecta($this->ProductosModel->totalProductos());
-        $this->addData('producto', $this->ProductosModel->getAllProducto($this->conf->getConfiguracion()->cantPagina,$_GET['offset']));
-        $this->addData('pag', $_GET['pag']);
-        $this->addData('method', "ProductosListar");
-        $this->display("ProductosListarTemplate.twig");
-    }
-    public function productosAM(){
-        if (isset($_POST["idProducto"])){
-            $this->addData('producto', $this->ProductosModel->searchIdProducto($_POST["idProducto"]));
+    public function productosListar()
+    {
+        if($this->permissions())
+        {
+            $this->paginaCorrecta($this->ProductosModel->totalProductos());
+            $this->addData('producto', $this->ProductosModel->getAllProducto($this->conf->getConfiguracion()->cantPagina,$_GET['offset']));
+            $this->addData('pag', $_GET['pag']);
+            $this->addData('method', "ProductosListar");
+            $this->display("ProductosListarTemplate.twig");
         }
-        $this->addData('categoria',$this ->CategoriaModel->getAllCategorias());
-        $this->display("ProductosAMTemplate.twig");
     }
-    public function productosAMPost(){
-        try{
-            $this->validarProductos($_POST);
-            if (isset($_POST['idProducto'])){
-                $this->addData('producto',$this ->ProductosModel->actualizarProducto($_POST));
-            }else{
-                $this->addData('producto',$this ->ProductosModel->insertarProducto($_POST));
+    public function productosAM()
+    {
+        if($this->permissions())
+        {
+            if (isset($_POST["idProducto"])) $this->addData('producto', $this->ProductosModel->searchIdProducto($_POST["idProducto"]));
+
+            $this->addData('categoria', $this->CategoriaModel->getAllCategorias());
+            $this->display("ProductosAMTemplate.twig");
+        }
+    }
+
+    public function productosAMPost()
+    {
+        if($this->permissions()){
+            try
+            {
+                $this->validarProductos($_POST);
+
+                if (isset($_POST['idProducto'])) $this->addData('producto',$this ->ProductosModel->actualizarProducto($_POST));
+                else $this->addData('producto',$this ->ProductosModel->insertarProducto($_POST));
+
+                $_GET['pag'] = 0;
+                $this->addData('method', "ProductosListar");
+                $this->productosListar();
+
             }
+            catch (Exception $e)
+            {
+                $this->addData('producto',$_POST);
+                $this->addData('mensajeError', $e->getMessage());
+                $this->productosAM();
+            }
+        }
+    }
+
+    public function productosEliminar()
+    {
+        if($this->permissions()) {
+            $this->addData('producto', $this->ProductosModel->deleteProducto($_POST["idProducto"]));
             $_GET['pag'] = 0;
             $this->addData('method', "ProductosListar");
             $this->productosListar();
-        } catch (Exception $e){
-            $this->addData('producto',$_POST);
-            $this->addData('mensajeError', $e->getMessage());
-            $this->productosAM();
         }
     }
 
-    public function productosEliminar(){
-        $this->addData('producto', $this ->ProductosModel->deleteProducto($_POST["idProducto"]));
-        $_GET['pag'] = 0;
-        $this->addData('method', "ProductosListar");
-        $this->productosListar();
-    }
-
-    public function validarProductos($var){
+    public function validarProductos($var)
+    {
         $this->validator->varSet($var['submit'],"Presione el boton de submit");
         $this->validator->validarStringEspeciales($var['nombre'],"Error: el campo 'Nombre' solo admite letras.",25);
         $this->validator->validarStringEspeciales($var['marca'],"Error: el campo 'Marca' solo admite letras.",25);
@@ -67,12 +86,13 @@ class ProductoController extends Controller
         $this->validator->validarNumeros($var['stockMinimo'],"Error: el 'Stock Minimo' nombre solo admite letras.",3);
         $this->validator->validarNumeros($var['idCategoria'],"No existe la categoria.",3);
         $this->validator->validarNumerosPunto($var['precioVentaUnitario'],"Error: solo se admite el formato xxx.xx",5);
-        if (! $this->CategoriaModel->getCategoriaById($var['idCategoria'])) throw new valException("Error:No existe la categoria");
+        if (! $this->CategoriaModel->getCategoriaById($var['idCategoria'])) throw new Exception("Error:No existe la categoria");
     }
 
 
     /*--- paginacion ---*/
-    public function paginaCorrecta($total){
+    public function paginaCorrecta($total)
+    {
         if (! isset($_GET['pag'])) throw new Exception('Error:No hay una página que mostrar');
         elseif ($total->total <= $_GET['pag'] *  $this->conf->getConfiguracion()->cantPagina){  $_GET['pag'] = 0; $_GET['offset'] = 0;}
         else $_GET['offset'] = $this->conf->getConfiguracion()->cantPagina * $_GET['pag'];
